feat(mobile/shlokas): promote Shlokas tab into a Sacred Hub
Slot 3 of the bottom bar becomes the single doorway to every sacred
instrument in Kiaanverse, matching the design reference. One tap from
Shlokas reaches:

  Sacred Scriptures:  Bhagavad Gita (18 chapters + daily verse)
  Healing Tools:      Emotional Reset, Karma Reset
  Wisdom Tools:       Ardha, Viyoga, Relationship Compass
  Sacred Community:   Wisdom Room, Sacred Reflections
  Sacred Sound:       KIAAN Vibe Player (full-width card)

Changes:

- `(tabs)/shlokas/index.tsx` — new Sacred Hub screen composed from the
  existing ToolCard / SectionHeader / VibePlayerCard primitives, with
  the same lotus-bloom entrance choreography used by `/tools`. Sections
  stagger by 100 ms, cards by 60 ms each.
- `(tabs)/shlokas/gita.tsx` — the former `index.tsx` (Bhagavad Gita
  browser) renamed so the tab root can host the hub. Existing deep-link
  routes `/shlokas/[chapter]/[verse]` stay intact — Daily-Verse banners
  and push-notifications still resolve without edits.
- `(tabs)/shlokas/_layout.tsx` — registers the new `gita` stack screen
  alongside `index` and the `[chapter]/[verse]` routes.
- `(tabs)/_layout.tsx` — un-hides `shlokas` (slot 3) and hides `journeys`
  (still routable; reached via Home's "Browse Sacred Catalog" CTA and
  push-notifications). Keeps the bar at five doorways.
- `components/navigation/DivineTabBar.tsx` — TABS array now maps slot 3
  → shlokas/ManuscriptIcon (scroll of shlokas) and slot 4 →
  journal/ChakraColumnIcon (inward ascent), matching the screenshot.

Non-changes:
- `/tools/index.tsx` is untouched. It stays as an alternate surface for
  any existing callers/deep-links and intentionally mirrors a subset of
  the hub's sections.
- No native config touched → this ships as a pure OTA update via the
  new mobile-ota-update.yml workflow on merge to main.
