+++ kiaanverse-mobile/apps/mobile/app/(tabs)/_layout.tsx
@@ -1,12 +1,18 @@
 /**
  * Tabs Layout — 5-tab bottom navigation for Kiaanverse.
  *
- *   Home · Sakha · Journeys · Journal · Profile
+ *   Home · Sakha · Shlokas · Journal · Profile
  *
- * Shlokas used to live in slot 3 and Journeys was buried as a sub-tab of
- * Journal; the layout promotes Journeys to a top-level destination and
- * keeps Shlokas reachable via deep-links (from the Vibe Player's Daily
- * Verse banner, notifications, etc.) while removing it from the tab bar.
+ * Shlokas owns slot 3 as a Sacred Hub — every sacred instrument the user
+ * has (Gita scriptures, healing + wisdom tools, Wisdom Rooms, Sacred
+ * Reflections, KIAAN Vibe Player) is one tap away from this tab's root
+ * screen. The full shloka browser lives at `/shlokas/gita`, with existing
+ * deep-link routes `/shlokas/[chapter]/[verse]` untouched.
+ *
+ * Journeys is still a first-class route (reachable from Home's "Browse
+ * Sacred Catalog" CTA and from notifications), but it is `href: null`'d
+ * here so the bottom bar stays at five doorways and the Shlokas Hub is
+ * the home for exploration.
  *
  * - Uses the custom DivineTabBar (gold-accented, dark navy background).
  * - `freezeOnBlur` preserves tab state (scroll positions, chat history,
@@ -14,9 +20,6 @@
  * - Deep-link routes (`/journey/...`, `/verse/...`, `/journal/new`, etc.)
  *   live outside the (tabs) group and are pushed with the Expo Router
  *   stack, not rendered in the tab bar.
- * - `shlokas` is registered with `href: null` so its route is still valid
- *   (every existing `router.push('/(tabs)/shlokas/...')` call in the app
- *   keeps working) but it doesn't render in the bottom bar.
  */
 
 import React from 'react';
@@ -38,13 +41,13 @@ export default function TabsLayout(): React.JSX.Element {
     >
       <Tabs.Screen name="index" options={{ title: t('home') }} />
       <Tabs.Screen name="chat" options={{ title: t('chat') }} />
-      <Tabs.Screen name="journeys" options={{ title: t('journeys') }} />
+      <Tabs.Screen name="shlokas" options={{ title: t('shlokas') }} />
       <Tabs.Screen name="journal" options={{ title: t('journal') }} />
       <Tabs.Screen name="profile" options={{ title: t('profile') }} />
-      {/* Hidden from the bottom bar but still routable — see header note. */}
+      {/* Routable but hidden from the bar — reached from Home's CTA. */}
       <Tabs.Screen
-        name="shlokas"
-        options={{ title: t('shlokas'), href: null }}
+        name="journeys"
+        options={{ title: t('journeys'), href: null }}
       />
     </Tabs>
   );
